+++ src/test/java/com/island/ReproducibilityTest.java
@@ -5,6 +5,7 @@ import com.island.engine.event.DefaultEventBus;
 import com.island.nature.config.Configuration;
 import com.island.nature.model.DefaultBiomassManager;
 import com.island.nature.model.Island;
+import com.island.nature.model.StaticChunkingStrategy;
 import com.island.nature.service.AnimalHealthSystem;
 import com.island.nature.service.AnimalMovementSystem;
 import com.island.nature.service.DefaultProtectionService;
@@ -74,6 +75,7 @@ class ReproducibilityTest {
                 .biomassManager(new DefaultBiomassManager())
                 .randomProvider(fixedProvider)
                 .componentRegistry(componentRegistry)
+                .chunkingStrategy(new StaticChunkingStrategy(config))
                 .build();
 
         Island island = new Island(context, 2, 2, new DefaultEventBus());