+++ src/test/java/com/island/nature/entities/WolfPackBalanceTest.java
@@ -7,6 +7,7 @@ import com.island.nature.entities.predators.Bear;
 import com.island.nature.model.Cell;
 import com.island.nature.model.DefaultBiomassManager;
 import com.island.nature.model.Island;
+import com.island.nature.model.StaticChunkingStrategy;
 import com.island.nature.service.AlertService;
 import com.island.nature.service.DefaultProtectionService;
 import com.island.nature.service.AnimalFeedingSystem;
@@ -66,6 +67,7 @@ class WolfPackBalanceTest {
                 .alertService(new AlertService())
                 .protectionService(new DefaultProtectionService(config, registry, statisticsService, 1))
                 .biomassManager(new DefaultBiomassManager())
+                .chunkingStrategy(new StaticChunkingStrategy(config))
                 .randomProvider(randomProvider)
                 .componentRegistry(componentRegistry)
                 .build();