+++ src/test/java/com/island/nature/service/AnimalMovementSystemTest.java
@@ -11,6 +11,7 @@ import com.island.nature.entities.registry.SpeciesRegistry;
 import com.island.nature.model.Cell;
 import com.island.nature.model.DefaultBiomassManager;
 import com.island.nature.model.Island;
+import com.island.nature.model.StaticChunkingStrategy;
 import com.island.util.common.DefaultRandomProvider;
 import com.island.util.interaction.InteractionMatrix;
 import com.island.util.math.GridUtils;
@@ -50,6 +51,7 @@ class AnimalMovementSystemTest {
                 .animalFactory(animalFactory)
                 .protectionService(new DefaultProtectionService(config, registry, stats, 4))
                 .biomassManager(new DefaultBiomassManager())
+                .chunkingStrategy(new StaticChunkingStrategy(config))
                 .randomProvider(random)
                 .componentRegistry(componentRegistry)
                 .build();